+++ src.mk
@@ -163,6 +163,7 @@ TEST_BENCH_SOURCES =                                                    \
   db/dbformat_test.cc                                                   \
   db/db_iter_test.cc                                                    \
   db/db_test.cc                                                         \
+  db/db_compaction_filter_test.cc                                       \
   db/db_dynamic_level_test.cc                                           \
   db/deletefile_test.cc                                                 \
   db/fault_injection_test.cc                                            \
